auth api fix
