+++ src/store/actions/sessionAction.js
@@ -1,5 +1,6 @@
 import actionTypes from '../constants';
-import sessionServices from '../services/sessionServices';
+// import sessionServices from '../services/sessionServices';
+import Api from '../services/api';
 
 const sessionUser = (path) => async (dispatch) => {
   dispatch({
@@ -7,7 +8,19 @@ const sessionUser = (path) => async (dispatch) => {
   });
 
   try {
-    const user = await sessionServices(path);
+    // const user = await sessionServices(path);
+    const user = await Api(path, 'post')
+      .then((response) => {
+        console.log(response);
+        const { data } = response;
+        if (data.error) {
+          throw data.error;
+        }
+        return data;
+      })
+      .catch((error) => {
+        throw error;
+      });
     dispatch({
       type: actionTypes.SESSION_SUCCESS,
       payload: {
